pyproject.toml
[tool.poetry]
name = "dataverse"
version = "0.1.0"
description = ""
authors = ["romannocry <roman.medioni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.39"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"
sqlmodel = "^0.0.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
